# [ENH] Perf for left/right join when `sort_by_appearance` is False